Remove distutils dependency
distutils was removed in Python 3.12, per deprecation in PEP
632. It is not present in Debian Trixie and higher.

Removing distutils from Kicksecure dependencies fixes issues
when distro-morphing Debian Trixie and higher to Kicksecure. It
is also necessary for installing Kicksecure on ports-only
architectures such as ppc64.

As best I can tell, setup-wizard-dist is not actually using
distutils anywhere, so no code changes or new dependencies
should be necessary. If I am mistaken on this,
python3-setuptools is supposedly a drop-in replacement.